Corrected compiler warnings

+++ include/tudat/simulation/estimation_setup/createDopplerPartials.h
@@ -327,8 +327,8 @@ createTwoWayDopplerPartials(
         if( currentTwoWayDopplerPartial != nullptr )
         {
             // Add partial to the list.
-            std::pair< double, double > currentPair = std::pair< int, int >(
-                    parameterIterator->first, parameterIterator->second->getParameterSize( ) );
+            std::pair< double, double > currentPair = std::pair< double, double >(
+                static_cast< double >( parameterIterator->first ), static_cast< double >( parameterIterator->second->getParameterSize( ) ) );
             twoWayDopplerPartialList[ currentPair ] = currentTwoWayDopplerPartial;
         }
     }
